move current input_stream* files to their own directory
- add src/input/input_stream_factory to export them to the rest of the
  code
