+++ src/input/input_stream_factory.ts
@@ -1,0 +1,5 @@
+import BrowserInputStream from './input_stream/input_stream_browser';
+import NodeInputStream from './input_stream/input_stream_node';
+
+export default BrowserInputStream;
+export { NodeInputStream };
